Rename some database columns to make them fit a convention
Label column should be used for human readable names that are going to be displayed,
name column should be used for machine readable names that are used within the code,
in the field_types table, values in the rails_type column should match the name of a function in
http://api.rubyonrails.org/classes/ActionView/Helpers/FormTagHelper.html
without the _tag part. E.g. text_field or telephone_field or text_area.