Change the way we list items, we now show fields that should be shown
To decide which fields are shown, there is a new column in the form_fields table, named display
proper values should be NULL (not shown), show (to show in the table), and label, which only one field
for each form should have. This will ultimately be used where we can only show one field to identify a form
(e.g. in a breadcrumb)

Also discovered the find_or_create_by function, which allows some templates to be simplified